Move MMTK off-by-one fixup into mutator
The existence of foreign calls meant it was not possible to reliably
ensure that the mutator would not encounter a null cursor as MMTk resets
all nurseries when garbage collecting.  This meant that threads blocked
in foreign calls would underflow their Hp when applying the off-by-one
fixup upon reopening their nursery.

To avoid this we instead apply the off-by-one fixup only if the nursery
has been initialized.